Extract common complexity report logic to metrics module
This allows to share complexity report logic between the console report and a report-html module.
